+++ lib/yabeda/activerecord.rb
@@ -1,10 +1,3 @@
 # frozen_string_literal: true
 
-require_relative "activerecord/version"
-
-module Yabeda
-  module Activerecord
-    class Error < StandardError; end
-    # Your code goes here...
-  end
-end
+require_relative "active_record"