+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## [Unreleased]
+## Unreleased
 
-## [0.1.0] - 2022-05-26
+## 0.1.0 - 2022-05-27
 
-- Initial release
+- Initial release: query performance metrics and connection pool stats. [@Envek][]
+
+[@Envek]: https://github.com/Envek/ "Andrey Novikov"